requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9.9
playwright>=1.52.0
python-dotenv>=1.0.1
Pillow>=10.2.0
requests>=2.31.0
